requirements.txt
Flask==3.1.1
numpy==2.3.1
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.10.0
Werkzeug==3.1.0
Jinja2==3.1.5
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask-Bcrypt
Flask-Dance
requests
email-validator
Flask-Mail
itsdangerous
